epic refactor
добавил недостающие requiremnts
разбил на модули
отступы и название классов
