+++ app/config/config.py
@@ -1,0 +1,5 @@
+# rinat
+postgresqlConfig = "postgresql://postgres:rinat@localhost:5432/postgres"
+# serega
+# postgresqlConfig = "postgresql://postgres:1@localhost:5432/postgres"
+
