+++ app/models/surveys.py
@@ -1,46 +1,9 @@
 from datetime import datetime
-from flask_sqlalchemy import SQLAlchemy
+from app.config.db import db
 from sqlalchemy.dialects.postgresql import JSONB, ARRAY
 
-db = SQLAlchemy()
 
-class tokenblocklist(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    jti = db.Column(db.String(36), nullable=False, index=True)
-
-class users(db.Model):
-    def __init__(self, login, hash_password, email, role):
-        self.login = login
-        self.hash_password = hash_password
-        self.email = email
-        self.role = role
-
-    id = db.Column(db.Integer, primary_key=True)
-    login = db.Column(db.String(80), unique=True, nullable=False)
-    hash_password = db.Column(db.String(), nullable=False)
-    email = db.Column(db.String(120), nullable=False)
-    role = db.Column(db.String(1), nullable=False)
-    survey = db.relationship("surveys", backref="users")
-    profile = db.relationship("profiles", backref="users")
-
-    @classmethod
-    def get_by_id(self, id):
-        return id
-
-class profiles(db.Model):
-    def __init__(self, username, avatar_url, user_id):
-        self.username = username
-        self.avatar_url = avatar_url
-        self.user_id = user_id
-
-    id = db.Column(db.Integer, primary_key=True)
-    username = db.Column(db.String(), nullable=False)
-    avatar_url = db.Column(db.String(), nullable=True)
-    balance = db.Column(db.Integer, default=0)
-    complete_survey = db.Column(db.Integer, nullable=True, default=0)
-    user_id = db.Column(db.Integer, db.ForeignKey("users.id"))
-
-class surveys(db.Model):
+class Surveys(db.Model):
     def __init__(self, title, description, logoPosition, value, pages, user_id):
         self.title = title
         self.description = description
@@ -57,9 +20,10 @@ class surveys(db.Model):
     date_creation = db.Column(db.DateTime(), nullable=False, default=datetime.utcnow)
     pages = db.Column(ARRAY(JSONB), nullable=True)
     user_id = db.Column(db.Integer, db.ForeignKey('users.id'))
-    page = db.relationship("pages", backref="surveys")
+    page = db.relationship("Pages", backref="surveys")
 
-class pages(db.Model):
+
+class Pages(db.Model):
     def __init__(self, name, elements, surveys_id):
         self.name = name
         self.elements = elements
@@ -69,9 +33,10 @@ class pages(db.Model):
     name = db.Column(db.String(), nullable=False)
     elements = db.Column(ARRAY(JSONB), nullable=False)
     surveys_id = db.Column(db.Integer, db.ForeignKey("surveys.id"))
-    question = db.relationship("questions", backref="pages")
+    question = db.relationship("Questions", backref="pages")
 
-class questions(db.Model):
+
+class Questions(db.Model):
     def __init__(self, type, name, isRequired, title, placeholder, choices, page_id):
         self.type = type
         self.name = name
@@ -89,13 +54,14 @@ class questions(db.Model):
     placeholder = db.Column(db.String(), nullable=True)
     choices = db.Column(ARRAY(db.String()), nullable=True)
     page_id = db.Column(db.Integer, db.ForeignKey("pages.id"))
-    answer = db.relationship("answers", backref="questions")
+    answer = db.relationship("Answers", backref="questions")
 
     def serialize(self):
         fields = ["_sa_instance_state", "id", "page_id"]
         return {k: v for k, v in self.__dict__.items() if v != None and k not in fields}
 
-class answers(db.Model):
+
+class Answers(db.Model):
     def __init__(self, title, answer, question_id, user_id):
         self.title = title
         self.answer = answer
@@ -107,5 +73,3 @@ class answers(db.Model):
     answer = db.Column(db.String(), nullable=False)
     question_id = db.Column(db.Integer, db.ForeignKey("questions.id"))
     user_id = db.Column(db.Integer, db.ForeignKey("users.id"))
-
-
