+++ run.py
@@ -1,0 +1,4 @@
+from app.app import app
+
+if __name__ == "__main__":
+    app.run(debug=True, port=3000, host='127.0.0.1')
