--- templates/base.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content = "width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>World</title>
-    <link rel="stylesheet" href="./static/css/style.css">
-</head>
-<body>
-    <main>
-        <h1>{% block content %} {% endblock %}</h1>
-    </main>
-</body>
-</html>
