--- templates/index.html
@@ -1,6 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-    <h1>Мир посетили: {{visited}}</h1></p>
-
-{% endblock %}
